pyproject.toml
[project]
name = "tagprocess"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gevent>=25.5.1",
    "pymysql>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
]
